requirements.txt
aiogram==3.*
python-dotenv
Pillow
pytest
pytest-asyncio
coverage